+++ src/main/java/com/buuz135/functionalstorage/block/tile/CompactingDrawerTile.java
@@ -39,6 +39,7 @@ public class CompactingDrawerTile extends ItemControllableDrawerTile<CompactingD
             @Override
             public void onChange() {
                 CompactingDrawerTile.this.markForUpdate();
+                CompactingDrawerTile.this.updateComparatorOutput();
             }
 
             @Override