+++ src/main/java/com/buuz135/functionalstorage/block/tile/DrawerTile.java
@@ -35,6 +35,7 @@ public class DrawerTile extends ItemControllableDrawerTile<DrawerTile> {
             @Override
             public void onChange() {
                 DrawerTile.this.markForUpdate();
+                DrawerTile.this.updateComparatorOutput();
             }
 
             @Override