+++ src/main/java/com/buuz135/functionalstorage/block/tile/ItemControllableDrawerTile.java
@@ -140,6 +140,7 @@ public abstract class ItemControllableDrawerTile<T extends ItemControllableDrawe
                 })
                 .setOnSlotChanged((stack, integer) -> {
                     setNeedsUpgradeCache(true);
+                    updateComparatorOutput();
                 })
                 .setSlotLimit(1);
     }