+++ src/main/java/com/buuz135/functionalstorage/block/tile/SimpleCompactingDrawerTile.java
@@ -39,6 +39,7 @@ public class SimpleCompactingDrawerTile extends ItemControllableDrawerTile<Simpl
             @Override
             public void onChange() {
                 SimpleCompactingDrawerTile.this.markForUpdate();
+                SimpleCompactingDrawerTile.this.updateComparatorOutput();
             }
 
             @Override